refactor: combine createDocClient, updateDocAdmin onSubmit
