+++ apps/web/src/components/filling-detail/create-edit/createDocumentClient.tsx
@@ -18,13 +18,10 @@ import ButtonPanel from './buttonPanel';
 import FileInputPanel from './fileInputPanel';
 import ActivityPanel from './activityPanel';
 import { DocumentType } from '@/src/interface/document';
-import createDocument from '@/src/service/document/createDocument';
-import { DocumentActivity, FilingStatus } from '@/src/constant/enum';
-import { getFileType } from '@/src/lib/utils';
-import updateFilingName from '@/src/service/filing/updateFiling';
+import { FilingStatus } from '@/src/constant/enum';
 import { toast } from '../../ui/use-toast';
-import uploadFileToS3 from '@/src/service/aws/uploadFileToS3';
-import { zodDocumentFiles } from '@/src/constant/schema';
+import { createdFormSchema } from '@/src/constant/schema';
+import submitCreatedFormSchema from '@/src/lib/submitCreatedFormSchema';
 
 export default function CreateDocumentClient({
   setShowCreateDocument,
@@ -39,14 +36,6 @@ export default function CreateDocumentClient({
   projectId: string;
   status: FilingStatus;
 }) {
-  const createdFormSchema = z.object({
-    // Server side ไม่รู้จัก FileList ***
-    file: zodDocumentFiles,
-    activity: z.nativeEnum(DocumentActivity, { message: 'กรุณากรอกกิจกรรม' }),
-    detail: z.string().min(1, { message: 'กรุณากรอกรายละเอียด' }),
-    note: z.string().optional(),
-  });
-
   const form = useForm<z.infer<typeof createdFormSchema>>({
     resolver: zodResolver(createdFormSchema),
     defaultValues: {
@@ -59,44 +48,14 @@ export default function CreateDocumentClient({
   async function onSubmit(values: z.infer<typeof createdFormSchema>) {
     // TODO: change to actual userId
     try {
-      const swap = getFileType(values.file[0]) !== 'pdf';
-      const pdfFile = values.file[swap ? 1 : 0];
-      const docFile = values.file[swap ? 0 : 1];
-      const folderName = `${projectId}/${filingId}`;
+      const newDocument = await submitCreatedFormSchema(
+        values,
+        projectId,
+        filingId,
+        'd1c0d106-1a4a-4729-9033-1b2b2d52e98a',
+        status,
+      );
 
-      const [pdfName, docName] = await Promise.all([
-        uploadFileToS3({
-          file: pdfFile,
-          folderName,
-        }),
-        docFile &&
-          uploadFileToS3({
-            file: docFile,
-            folderName,
-          }),
-      ]);
-
-      if (!pdfName || (docFile && !docName))
-        throw new Error('Upload file failed');
-
-      const [newDocument, _] = await Promise.all([
-        createDocument({
-          document: {
-            name: values.detail,
-            filingId,
-            pdfName: pdfName,
-            docName: docName ?? '',
-            activity: values.activity as DocumentActivity,
-            userId: 'd1c0d106-1a4a-4729-9033-1b2b2d52e98a',
-            detail: values.note,
-          },
-        }),
-        status === FilingStatus.DRAFT &&
-          updateFilingName({
-            filingId,
-            filingStatus: FilingStatus.DOCUMENT_CREATED,
-          }),
-      ]);
       afterCreateDocument(newDocument);
       toast({
         title: 'สร้างเอกสารสำเร็จ',
